requirements.txt
alembic==1.17.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.120.1
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic-settings==2.3.4
pydantic-extra-types==2.9.2
pydantic_core==2.41.4
python-dotenv==1.2.1
python-jose==3.3.0
cryptography==42.0.0
passlib==1.7.4
